created test_linops file. added operator(i,j) to SymmetricLinearOperator, ExplicitSymLinOp, and RegExplicitSymLinOp
